[3/n] [sled-agent] put zone image resolver behind Arc/Mutex (#8080)
The `ZoneImageSourceResolver` is going to gain the ability to reset the
mupdate override file, which involves synchronizing on-disk and
in-memory state. It's easiest to ensure that accesses to the resolver
are serialized.

We're also going to have to clone this resolver in the future, so use
`Arc`.

Since there's no async involved here, it's simpler to do this than use
the actor model.