[3/n] [installinator] write out mupdate-override.json (#8042)
Write out a file containing a UUID which indicates that installinator
wrote out an updated dataset. This is a signal for Nexus to assume that
installinator wrote out a new install dataset behind its back.

I'm not entirely sure if writing out the JSON file at the beginning vs
at some later point matters a huge amount here -- if installinator fails
for some reason, we assume the system is broken anyway.

We also record the host phase 2 and control plane artifact hashes, and
there's space to record other hashes if we'd like to do so in the
future.

Tested on dublin:

![Screenshot 2025-04-24
150911](https://github.com/user-attachments/assets/19ad609f-c3bd-423d-aac4-32e3bdfbf29e)

```
BRM42220026 # ls -al /pool/int/*/install/mupdate-override.json
-rw-------   1 root     root          55 Dec 28  1986 /pool/int/2196ba5d-e211-4495-9c0b-e0948ba0afbd/install/mupdate-override.json
-rw-------   1 root     root          55 Dec 28  1986 /pool/int/f3673b2a-0cdc-4315-ac53-7889918cee4c/install/mupdate-override.json
```

See [RFD 556] for more information.

[RFD 556]: https://rfd.shared.oxide.computer/rfd/0556