[sled-agent-config-reconciler] Flesh out internal disks task (#8103)
Nothing too fancy here. A lot of the bulk of the code is around
funneling results out through watch channels.

The only new behavior beyond what `sled-storage` does is that this will
periodically retry calling `Disk::new()` on internal disks for which
`Disk::new()` previously failed. I'm very open to feedback here - maybe
this is wrong because the most common errors are fatal anyway (e.g.,
disk isn't formatted correctly)? Maybe it's okay but should be less
frequent? Maybe it's okay but we should check for specific kinds of
errors we believe to be fatal and not retry in those cases?