deflake instance_failed_by_stop_request_does_not_reincarnate (#8152)
in this test the fact that the instance failed restarts the reincarnation task. so the typical case is actually that we're racing with that activation to detect the failed instance, then have a no-op reactivation and wait for the prior reactivation to restart the failed instance.

just remove the check for a Failed state and background task activation.

see #8119. or in more automatic issue management words:

fixes #8119.