[sled-agent-config-reconciler] Flesh out zone start/stop (#8140)
Adds an internal `OmicronZones` type that is analogous to the `ZoneMap`
in the current service manager. There are a couple of behavioral
differences between this and what ServiceManager does:

* Starting and stopping zones are broken up into separate methods (so
the reconciler can do disk/dataset work in between stopping zones and
starting zones)
* ServiceManager currently tries once to shut down a zone and logs any
errors, but otherwise just pretends it succeeds. (Discussion on #6776,
which largely ends with "we should do a config reconciler" :joy:.) This
attempts to remember failures to shut a zone down (or clean up after it)
and retry them later. I'm not sure all of this is right, or whether it
might even introduce some new kind of failure mode (e.g., a zone getting
stuck in some partially shut down state that we previously would have
ignored?).