Add protections for password handling in Nexus (#8093)
This wraps uses of user passwords in `secrecy` types that:

1. make an effort to not accidentally expose passwords (via Debug,
serialization, etc.)
2. zeroize memory when a password is no longer used.

The Nexus external API parameter type for passwords kept a local String
copy of user passwords in addition to the internal password type, the
former of which was used only for testing purposes. This removes that,
ensuring that passwords in Nexus are always wrapped in `secrecy`, but it
has the unfortunate requirement now that test code needs to use slightly
different parameter types than those exposed by the external Nexus API.
I've tried to make it clear where that happens in code and why in the
code comments, but happy to take suggestions on different approaches.