[sled-agent-config-reconciler] Add `DumpSetupTask` (#8146)
Adds a tokio task responsible for activating `DumpSetup` whenever the
set of disks changes. This is analogous to the current
[`StorageMonitor`](https://github.com/oxidecomputer/omicron/blob/3f5c3f31cc26ca35448141f6da7cf8c56be65c3f/sled-agent/src/storage_monitor.rs#L87-L116),
with two mechanical changes:

* It no longer makes sense to wait for changes to `generation`, because
that covers the entire config (and usually will be bumped just for
changes to the zones; e.g., during updates). We instead consume channels
that contain the managed disks and keep our own "what was the last set
of disks we used" to decide when to forward a change notification on to
`update_dumpdev_setup`.
* We consume two channels due to the split of internal vs external disks
within the config reconciler.