Raise Nexus' VM memory limit to 1T (#8160)
fixes #7918. ish. it would be nice to not have a hardcoded limit here at
all. while one might think there _shouldn't_ be a real difference in
behavior from here up to about 16 TiB VMs, reality confounds.
https://github.com/oxidecomputer/propolis/issues/907 is the next issue
we'll run into for raising VM size limits further.

Angela has instances up to and including 1T a try with notes on #7918.
it would be nice to set this as high as 1.5T, since that fits well on a
Gimlet with 2T of memory, but trying that is how we found Propolis#907.