[external API] alerts: the renamening (#8169)
In conversation with @ahl, we have determined that the external API for
webhooks added in #7277 should be changed to focus on "alerts" as the
first-class user-facing concept, with "webhooks" as one delivery
mechanism for alerts. This way, we can talk about alerts as an entity in
the API that exist independently of webhooks that deliver alerts, and
the same alert types can be shared with other alert delivery mechanisms
if any are added in the future.

What we currently refer to as "webhook events" and "webhook event
classes" are therefore renamed to "alerts" and "alert classes". The
current concept of "webhook receivers" is generalized to an "alert
receiver" resource, of which webhook receivers are (currently) the only
subtype. This way, if we add other mechanisms of delivering alerts in
the future (email, first-class Slack integration, etc), we can introduce
new subtypes of alert receivers. I've restructured the API to have both
`/v1/alert-receivers/...` and `/v1/webhook-receivers/...` routes, with
operations common to all alert receivers (list, view, add/remove
subscriptions, delete) under the `alert-receivers` route, and operations
related to webhook-specific configuration (add/remove secrets, probe,
deliveries) under the `webhook-receivers` route. I've also changed the
`AlertReceiver` view to have a "kind" enum that stores the
subtype-specific configuration; currently, this will only ever be
"webhook", but I thought it was worth doing this now to make future
additions cause less breakage for API consumers.

This is, admittedly, a somewhat large diff, but fortunately, most of it
is just renaming stuff and moving it around. Reviewers can focus more or
less exclusively to the changes to the external API routes and models,
and maybe the database migrations. Any mistakes while renaming and
moving things around have already been caught by the Rust compiler. :)