[nexus] rename webhook alert headers
Somehow, while working on renaming "webhook events" to "alerts" in
PR #8169, I managed to overlook the names of the headers
`x-oxide-event-id` and `x-oxide-event-class` that we set on webhook
delivery
requests. These should probably also be renamed to `alert`, so I did
that. I also renamed `x-oxide-webhook-id` to `x-oxide-alert-id`, as that
seemed clearer.