[bundles] Add range requests to bundle download API (#8183)
This PR adds range request support to the `support_bundle_download`
endpoint in the internal and external API.
It adds a `RangeRequest` header to both of these endpoints, and forwards
this request to the underlying Sled Agent.

Additionally, this PR updates the Sled Agent API to identify that range
requests are supported. They **were** supported before this PR, but not
encoded into the openapi spec until now.

This PR also adds:
- An integration test for Nexus, validating that range requests work for
bundle downloading
- omdb support for range requests when downloading bundles, including an
option to resume a bundle that has only partially downloaded