correct other test issues now that it actually changes DNS config
on one hand: now that DNS servers are referenced by potentially two
different AAAA records, both of those records are potentially the target
of a SRV record. though, we don't have SRV records for the DNS
interface. this test had failed at first because we'd find a DNS
server's IP via the `ns1.` record, which means we'd miss that the same
zone was referenced by an AAAA record for the illumos zone UUID.

on the other hand: #[nexus_test] environments involve a mock of the
initial RSS environment construction? so now that the first blueprint
adds NS records, this mock RSS environment was out of date, and a test
that the first blueprint after "RSS" makes no change failed because the
"RSS" environment was wrong.