Ensure next-item queries look at gaps in both directions (#8215)
- We modified the next-item queries in #6551 to reduce memory
consumption. In previous versions, we constructed the entire candidate
range of next items and joined that with the target table, taking the
first entry that doesn't exist in the table. But CRDB materializes all
subqueries eagerly, thus the memory problems. The modified version used
a self-join, between the list of allocated items and those items
incremented by 1. This limited memory consumption to be linear in the
number of _allocated_ items, rather than the full candidate range, which
is generally much smaller. However, this new query isn't exactly the
same as the original: it only looks at gaps _larger_ than allocated
items, whereas the original query inspected the entire range. This
commit reworks this self-join query to inspect gaps both above and below
the target items. In doing so, I've also restructured the query to be
much simpler and more direct, and I think much more efficient. All
subqueries in it are limited to at most 1 element, which should consume
constant rather than linear memory.
- Fixes #8208.