remove extraneous parameter from DnsConfigBuilder zone-specific helpers
for the service-specific builder methods, the service name was expected
as an argument but its only value was to check it was the same name that
builder expected. the Clickhouse-related methods then used other names
when they would set up multiple records!

the only value of the `service_name` argument seems to be as defensive
programming, but it seems to primarily be a kind of redundant footgun -
if there were a copypaste error this would have caught, we still would
define a service as multiple zones on the same sled. this would also
error! we would still detect erroneous configurations.