[1/n] Add `id` column on tokens and sessions (#8137)
Closes #8139. Token IDs are used in #8227 and #8231 for token CRUD.

WIP. Everything compiles and existing tests work, but we are not yet
testing retrieval by ID and we need to think about authz checks on the
datastore functions for retrieving tokens and sessions by token string.

This PR would be a lot shorter if not for the fact that the `token`
column was the primary key on both tables, so adding the `id` requires a
bunch of migration steps to change the primary key. There were also a
lot of changes to code and tests around making things ID-centric instead
of token-centric.

* [x] SQL migrations for adding `id` col and changing primary key to
that
* [x] Add `TypedUuid` to session and token DB models
* [x] Update `authz_resource!` and `lookup_resource!` calls for new
primary key
* [x] Update `Session` trait methods to use `id` instead of `token`
* [x] Update app and datastore session and token methods to use ID
instead of token
* [x] Update two distinct but nearly identical `FakeSession`
implementations (one for unit tests, one for integration tests) to a)
track sessions in a `Vec` instead of a `HashMap` with token keys
* [x] Figure out authz situation in new fetch session/token by token
string methods