[3/n] `/v1/me/access-tokens` list and delete  (#8227)
Built on top of #8137 and #8214.

This is only for a user to list and delete their own tokens. It doesn't
quite match RFD 570, which says `/v1/device-tokens` instead of
`/v1/me/access-tokens`, but it feels good under `/v1/me`, and after
trying to make the UI too, I think "access tokens" is much more
intuitive. If I stick with this, I will update RFD 570 to match.

~~I'm not sure about the path `/v1/device-tokens` — in the API we call
them `Device Access Tokens`. I think `/v1/access-tokens` might be more
intuitive because the `device` is sort of an implementation detail, it
refers to the OAuth device auth flow, which we are using. In practice,
the user just gets a token with the CLI and pastes a code into the web
UI and they don't have to think too much about it, so exposing that
detail in the name might not be worth it.~~ Went with
`/v1/me/access-tokens`.

- [x] Basic token list and delete
- [x] Basic integration tests
- [x] Finalize endpoint paths
- [x] Figure out authz story
  - Went with restricting datastore functions to current actor for now