[4/n] Let user request a specific TTL for a token (#8231)
Closes #8147.

Built on #8137, #8214, and #8227.

This is pretty straightforward, I think. The user gives us a TTL in
seconds at token request time. We store it on the request row. When they
come back in the later step to confirm the code and generate the token,
we retrieve the TTL, validate that it is less than the silo max (if one
is set), and we use it to generate the `time_expires` timestamp, which
cannot be changed later.

One slightly surprising bit is that we can't validate the TTL against
the silo max at initial request time because we don't have any idea what
silo the user is associated with until the confirm step. So probably
want to make sure we are handling TTL validation errors nicely in the
web console, because I think that's where they will show up.