Revert "DNS servers should have NS and SOA records (#8047)"
This reverts commit 3e68262c0d8a87ed54588549d13c245b91b3391c.

The change in 3e68262c0 does not handle upgrade sufficiently: when
internal DNS starts it tries to parse a `CurrentConfig` from a json blob
describing the previous server version's config. As Angela found on
dogfood, this means internal DNS will fail to start with an error about
"missing field `serial` at line 1 column ..."

We're reverting this to unblock dogfood, but will add this back in with
additional logic to handle this case.