nexus: clarify `InstanceCreate` disk attachment behavior (#8276)
Update the documentation for the `boot_disk` and `disks` attributes to
note that specifying a `boot_disk` counts against the disk attachment
limit.

When attempting to create an instance with both a `boot_disk` specified
and 8 `disks` specified, none of which are the boot disk, the API
returns the following error.

```
{
  "request_id": "8f62e5f8-c5c4-4d30-9436-47d331a59b1c",
  "error_code": "InvalidRequest",
  "message": "cannot attach more than 8 disks to instance"
}
```

This is confusing to readers since it's not obviously that `boot_disk`
counts against the disk attachment.