[dns-server] servers should be able to use older configs (#8278)
PR #8047 comes with an unfortunate upgrade-only bug: before an upgrade,
a system's DNS servers would write out a configuration without the new
`serial` field added in 8047. When upgraded, the DNS servers would then
try to load that config, see it is missing a `serial` field, and error
for every query.

We expect to replace the previous-format configuration immediately after
upgrade by regenerating a blueprint for the current system and executing
it. But we should be able to use the previous-format configuration
anyway, so that DNS functions enough to get a control plane capable of
planning and executing that blueprint.