[nix] update Nixpkgs to pick up Rust 1.87.0
Nothing to see here, folks, we just have to do this occasionally.

I may try to hack up the Nix flake to *not* provide the Rust toolchain
and instead just use normal, mutable Rustup so that we don't have to do this every time we bump the compiler. Bbut there may have been
some reason for doing it this way...