[sled-agent-config-reconciler] Add separate `OmicronDatasets` type (#8218)
This is more consistent with how the reconciler remembers zones and
disks. This is almost all just moving code around. The only nontrivial
changes are:

1. In the spot where we ought to delete datasets, we at least "forget"
them in-memory and log an error about leaking the ZFS dataset
2. The dataset serialization task remembers fewer details about datasets
it ensured (just the names - that's all we need for the error checking
it does when dealing with nested datasets)

Staged on top of #8064. PR 1 of 3 working towards #8173.