ls-apis: mention related RFDs and sections (#8324)
The existing comments in `api-manifest.toml` were helpful, but I'd had a hard
time determining the difference between "versioned_how", versioning mechanisms
described in `openapi-manager`, and the words in RFD 532. Additionally, 532 and 
`api-manifeset.toml` talk about "updated before", and RFD 565 says more about
the update sequence, so point to it as well.