[sled-agent] Destroy orphaned datasets (PR 3/2) (#8323)
This allows the config-reconciler to destroy durable
datasets it believes are orphaned due to expunged Omicron zones. This is
disabled by default, and can be enabled on a sled-by-sled case via a new
`omdb sled-agent chicken-switch destroy-orphans enable` subcommand (with
related commands to `get` and `disable` the same).

We don't want to ship automatic dataset deletion before R17 (R16 should
only ship "report orphaned datasets"), but we need to be able to turn it
on for upgrade testing in the meantime. All of this `chicken-switch`
stuff should be removeable after R16, once we're comfortable enabling
deletion in general.