Add a TLA+ model for trust quorum
This includes a configuration file for running TLC.

This model is meant to formalize what is written as prose in RFD 238.
However, it actually makes a few changes, specifically around config
retrieval and ordering of commits and the ability to rotate keys across
multiple configurations. These are improvements, and RFD 238 will be
updated to reflect these chagnes.