[blueprint-planner] Don't fail on sleds with no zpools
Fixes #7800. That issue notes that this is a minor inconvenience when
adding a sled, but (a) it's easy to fix and (b) we hit it in colo in a
different context (expunging all the disks on a sled). That combo made
it seem worth it to go ahead and fix it.