Trust Quorum: Better match protocol in TLA+ spec
This PR changes the message implementation a bit and the
`PersistentState` structure. It adds messages that will be used in
follow up PRs, and ensures that `rack_id` is always checked on each
message. Furthemore it stores configurations and shares separately
inside the persistent state. This allows retrieval of a configuration
for stale sled and the recomputation of its own share for that
configuration to occur in separate steps. This matches what is done in
the TLA+ spec.

I also added a comment about the intention of future code coming for
the `Node` type. This reflects the experience of implementing LRTQ, and
is helpful to simplify testing. If we don't have to track timeouts for
arbitrary API requests inside the sans-io code, it becomes much simpler.