[nexus] add tokio-dtrace probes
I've written [a crate][1] that adds USDT probes for Tokio runtime
events, using the unstable callbacks provided by the Tokio runtime. This
commit adds that crate to Nexus. Note that, because `tokio-dtrace`
releis on Tokio's unstable features, it was necessary to add config to
`.cargo/config.toml` to enable the `--cfg tokio_unstable` rustc config
when compiling for illumos. In addition, adding the code necessary
configure the Tokio runtime to record these probes required some changes
to `bin/nexus.rs`, to change the use of the `#[tokio::main]` attribute
to manual use of the runtime builder.

I'd like to add these DTrace
probes in other binaries in Omicron as well, but figured I'd start with
Nexus as a proof of concept. It might be worth some general refactoring
of our `main` functions in the process of doing so.

[1]: https://github.com/oxidecomputer/tokio-dtrace