[10/n] [sled-agent] validate zone images as written in zone manifest (#8190)
Included in this PR:

* Add validation for zone images
* Write tests for success and failure cases
* Extend the `test_installinator_fetch` wicketd integration test to also
ensure installinator + sled-agent have a coherent view of the world

The last point forced me to expose a bunch of the status via a public
interface. We'll be able to reuse this information (though probably not
in this full form) and send it up as part of the inventory.