Add omdb,db,nexus api support for chicken switches
This is an initial commit to support chicken switches that allow runtime
reconfigurator configuration.

There is only one initial chicken switch which can be used to enable
the planner background task to run.

A follow up PR will be made to add a background task
that reads the chicken switch DB table and makes it
available to nexus via a watcher. This PR will complete
https://github.com/oxidecomputer/omicron/issues/8253 .

Support for listing history was added at the datastore level, but not
made available yet in OMDB, because of the way pagination works. This
will likely come in a follow up.