[12/n] [sled-agent] don't start install dataset zones on zone manifest error (#8237)
For mupdate overrides, in order to be safe, we must know that the data
stored in the JSON is consistent with the images stored on disk. We
currently apply this logic to install dataset hashes, and will apply it
to artifact hashes in the future.

Questions:

* [x] ~~Should we apply this logic to more critical zones like the
switch zone? They're always going to be part of the ramdisk maybe?~~ We
unconditionally succeed for RAMdisk zones, and in particular the switch
zone, now.
* [ ] How does this interact with config-reconciler work? On error, that
would permanently be blocked it seems to me. Note: we decide errors
per-zone now.
* [ ] We ask for the boot zpool to be passed in but also cache it as
part of constructing the info... how do we reconcile the relative
presence or absence of this info?

Depends on #8235.