[1/5] User data export: DB models and queries
This set of PRs implements the endpoints proposed in RFD 563, and the
set starts with adding the database parts: the models, the schema
change, and the query routines.

A "user data export" object records the attachment of a read-only volume
to a particular Pantry for the purpose of exporting data. These objects
will be created and deleted using sagas invoked by a background task,
and not managed by a user or operator. As of right now the only
read-only user resources that can be exported are snapshots and images.

A user data export object will be created for each non-deleted snapshot
and image. A "changeset" representing the work required to manage this
will be queried for by the associated background task.

During Pantry expungement, that Pantry will no longer be returned by a
DNS lookup for all in-service Pantry addresses. A list of in-service
addresses returned like this functions similarly to a rendezvous table,
and is used to soft-delete user data export objects affected by
expungements so they can be recreated by the associated background task.