update zip to 4.2.0 (#8425)
zip 2.6.x was yanked, which causes r-a in Omicron to be somewhat sad
(see https://github.com/oxidecomputer/tufaceous/pull/31 for example
output).

Along with omicron, we also pull in updates to tufaceous and
support-bundle-viewer.

The versions were yanked due to breaking semver (and republished as zip
version 3 and 4), but we already made our code compatible with zip 2.6
(https://github.com/oxidecomputer/tufaceous/pull/25) so this upgrade
involves no code changes.