Use `oxide-tokio-rt` crate for common Tokio runtime config (#8412)
Currently, all the async Rust binaries in the Omicron workspace use the
`#[tokio::main]` attribute on an `async fn main()`. This creates a Tokio
runtime with the default configuration provided by that attribute, and
runs the contents of the `async` main function on that runtime.

We now have need to configure the Tokio runtime used by Omicron
differently than the defaults provided by the `#[tokio::main]` macro. In
particular, we want to do the following:
- Disable the LIFO slot optimization, which can cause severe latency
spikes in some cases (see

https://github.com/oxidecomputer/omicron/issues/8334#issuecomment-2993159283)
- When running on illumos, configure runtime callbacks to emit
`tokio-dtrace` probes (see #8364)

In order to make it easier to use these configurations in all production
binaries in Omicron and other projects, I've introduced a new
[`oxide-tokio-rt`
crate](https://github.com/oxidecomputer/oxide-tokio-rt). This crate
provides
functions that configure a runtime with our desired settings and runs a
future
on it. This can be used to replace `#[tokio::main]` fairly simply, which
I've
done for most binaries in the repo.

There are a handful of binaries I did *not* update to use
`omicron-runtime`. In particular, I didn't change the use of
`#[tokio::main]` in example binaries (in particular, `oximeter-producer`
and `update-engine`), as I didn't want to add an `omicron-runtime`
dependency to those crates just for example. In the case of
`oximeter-producer`, in particular, which is depended on by code outside
of the Omicron repo, that would force those binaries to compile
`omicron-runtime` unnecessarily. Similarly, I didn't make some of the
dev-tools
binaries which are used in xtasks use `oxide-tokio-rt`, because I didn't
want to introduce another dependency that has to be compiled before an
xtask can run, and that felt more important than adding DTrace probes to
those simple commands --- if something goes wrong with them, we can add
the proves later. The exception to this are the `mgs-dev` and
`omicron-dev` binaries, which run MGS and Nexus *in-process*; in those
cases, I felt it was useful to have the DTrace probes enabled when
running in development. Similarly, I also thought it was worthwhile to
use `omicron-runtime` in OMDB, as it's deployed on production systems
and would be harder to recompile with DTrace probes if we end up
wanting them.

Closes #8364