[1/5] User data export: DB models and queries
This set of PRs implements the endpoints and functionality proposed in
RFD 563, and the set starts with adding the database parts: the models,
the schema change, and the query routines.

A "user data export" object records the attachment of a read-only volume
to a particular Pantry for the purpose of exporting data. These objects
will be created and deleted using sagas invoked by a background task,
and not managed by a user or operator. As of right now the only
read-only user resources that can be exported are snapshots and images.

A user data export object will be created for each non-deleted snapshot
and image. A "changeset" representing the work required to manage this
will be queried for by the associated background task. If a snapshot or
image is deleted, the associated user data export object will
automatically be marked for deletion, and cleaned up by a saga.

In order not to create unbounded work, a similar pattern of using a
state plus "locking" with an operating saga id is used. Otherwise the
background task's periodic activation could initiate an unbounded amount
of sagas based on the computed changeset.

During Pantry expungement, that Pantry will no longer be returned by a
DNS lookup for all in-service Pantry addresses. The list of in-service
addresses returned from a DNS lookup functions similarly to a rendezvous
table, and is used to mark user data export objects affected by
expungements for deletion so they can be recreated by the associated
background task.

Even if a Pantry is not expunged, a restart of the service (or a bounce
of the sled) will cause all volume attachments to be lost. The user data
export logic handles this by detecting when this occurs and marking any
affected user data export as deleted so that another attachment can be
created. This will be seen in future commits of this set.