Blueprint PlanningInput: Fewer `Option`s (#8470)
This gets rid of two `Option<_>`s in the planning input:

* `TufRepoPolicy::description` changes from `Option<TufRepoDescription>`
to non-optional `TargetReleaseDescription`. The latter is isomorphic to
`Option`, but spells the `None` case as
`TargetReleaseDescription::Initial`.
* `PlanningInput::old_repo` is now a non-optional `TufRepoPolicy`. In
the case where `tuf_repo` is "the initial policy" (i.e.,
`TargetReleaseDescription::Initial`, `old_repo` is _also_ set to the
initial policy. In all other cases, it's "the previous policy".